pyproject.toml
[tool.poetry]
name = "gptlab"
version = "0.1.0"
description = ""
authors = ["Artur Świerc <aswierc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
kink = "^0.6.5"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.0"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
fastapi-events = "^0.8"
httpx = "^0.25.0"
mypy = "^0.991"
black = "^23.10.1"
brotli = "^1.1.0"
python-multipart = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

